fix issues pointed by gemini-code-assist
